Move the Android backend to C.
