+++ Backends/Graphics5/Vulkan/Sources/kinc/backend/graphics5/vulkanunit.c
@@ -48,6 +48,11 @@
 
 #endif
 
+#ifdef _DEBUG
+#define VALIDATE
+#endif
+
+
 #include <vulkan/vulkan.h>
 
 #include <assert.h>
