+++ Backends/System/Android/Sources/kinc/backend/androidunit.c
@@ -1,0 +1,5 @@
+#include "audio.c.h"
+#include "display.c.h"
+#include "system.c.h"
+#include "window.c.h"
+#include "video.c.h"
