+++ Backends/System/Android/Sources/kinc/backend/display.c.h
@@ -23,61 +23,51 @@ int kinc_primary_display(void) {
 	return 0;
 }
 
-/*Display* Display::get(int index) {
-    if (index > 0) {
-        return nullptr;
-    }
-	return &display;
-}*/
-
-
 static int width() {
 	JNIEnv *env;
 	JavaVM *vm = kinc_android_get_activity()->vm;
-    (*vm)->AttachCurrentThread(vm, &env, NULL);
+	(*vm)->AttachCurrentThread(vm, &env, NULL);
 	jclass koreActivityClass = kinc_android_find_class(env, "tech.kinc.KincActivity");
 	jmethodID koreActivityGetScreenDpi = (*env)->GetStaticMethodID(env, koreActivityClass, "getDisplayWidth", "()I");
 	int width = (*env)->CallStaticIntMethod(env, koreActivityClass, koreActivityGetScreenDpi);
-    (*vm)->DetachCurrentThread(vm);
+	(*vm)->DetachCurrentThread(vm);
 	return width;
 }
 
 static int height() {
-	JNIEnv* env;
-    JavaVM *vm = kinc_android_get_activity()->vm;
-    (*vm)->AttachCurrentThread(vm, &env, NULL);
+	JNIEnv *env;
+	JavaVM *vm = kinc_android_get_activity()->vm;
+	(*vm)->AttachCurrentThread(vm, &env, NULL);
 	jclass koreActivityClass = kinc_android_find_class(env, "tech.kinc.KincActivity");
 	jmethodID koreActivityGetScreenDpi = (*env)->GetStaticMethodID(env, koreActivityClass, "getDisplayHeight", "()I");
 	int height = (*env)->CallStaticIntMethod(env, koreActivityClass, koreActivityGetScreenDpi);
-    (*vm)->DetachCurrentThread(vm);
+	(*vm)->DetachCurrentThread(vm);
 	return height;
 }
 
 static int pixelsPerInch() {
-	JNIEnv* env;
-    JavaVM *vm = kinc_android_get_activity()->vm;
-    (*vm)->AttachCurrentThread(vm, &env, NULL);
+	JNIEnv *env;
+	JavaVM *vm = kinc_android_get_activity()->vm;
+	(*vm)->AttachCurrentThread(vm, &env, NULL);
 	jclass koreActivityClass = kinc_android_find_class(env, "tech.kinc.KincActivity");
 	jmethodID koreActivityGetScreenDpi = (*env)->GetStaticMethodID(env, koreActivityClass, "getScreenDpi", "()I");
 	int dpi = (*env)->CallStaticIntMethod(env, koreActivityClass, koreActivityGetScreenDpi);
-    (*vm)->DetachCurrentThread(vm);
+	(*vm)->DetachCurrentThread(vm);
 	return dpi;
 }
 
 static int refreshRate() {
-	JNIEnv* env;
-    JavaVM *vm = kinc_android_get_activity()->vm;
-    (*vm)->AttachCurrentThread(vm, &env, NULL);
+	JNIEnv *env;
+	JavaVM *vm = kinc_android_get_activity()->vm;
+	(*vm)->AttachCurrentThread(vm, &env, NULL);
 	jclass koreActivityClass = kinc_android_find_class(env, "tech.kinc.KincActivity");
 	jmethodID koreActivityGetScreenDpi = (*env)->GetStaticMethodID(env, koreActivityClass, "getRefreshRate", "()I");
 	int dpi = (*env)->CallStaticIntMethod(env, koreActivityClass, koreActivityGetScreenDpi);
-    (*vm)->DetachCurrentThread(vm);
+	(*vm)->DetachCurrentThread(vm);
 	return dpi;
 }
 
-void kinc_display_init() {
-
-}
+void kinc_display_init() {}
 
 kinc_display_mode_t kinc_display_available_mode(int display_index, int mode_index) {
 	kinc_display_mode_t mode;
