--- Backends/System/Android/Sources/kinc/backend/displaydata.h
@@ -1,1 +1,0 @@
-#pragma once
